BLOCKS:  Split PhysicalTopologySource internal classes into separate files
